requirements.txt
PySide6>=6.6.0
Pillow>=10.0.0
pillow-heif>=0.14.0
rawpy>=0.19.0
exifread>=3.0.0
geopy>=2.4.0
